Fix type of file_type_index_ from size_t to int to get rid of printf format warning.
BUG=none
TEST=none

Review URL: http://codereview.chromium.org/552253

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@37716 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 956d033ff82196efedc74533bff9749ee9b82141

+++ chrome/browser/views/select_file_dialog.cc
@@ -81,7 +81,7 @@ class SelectFileDialogImpl : public SelectFileDialog {
 
     // The index of the default selected file filter.
     // Note: This starts from 1, not 0.
-    size_t file_type_index_;
+    int file_type_index_;
 
     // Default extension to be added to file if user does not type one.
     FilePath::StringType default_extension_;
